pyproject.toml
[project]
name = "my-research-assistant"
version = "0.1.0"
description = "Agentic research assistant, focused on Arxiv papers"
readme = "README.md"
authors = [
    { name = "Jeff Fischer" }
]
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "llama-index>=0.13.0",
    "mcp>=1.12.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "reportlab>=4.4.3",
]

[project.scripts]
my-research-assistant = "my_research_assistant:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
